Chore: Deprecate meastro card in documentation
We can no longer just copy paste the generated swagger file because deprecated routes have been manually removed.

To avoid this, we need to do some work and maintain two separate swagger generations because platform and recurring (removed routes) must still be available when swagger is copied to the gateway definition.

The quick fix is to remove the tooling to avoid future confusion.